+++ source/components/localisation/useLocalisation.ts
@@ -6,23 +6,12 @@
 	whereas the translation is about the text displayed to the user.
 */
 
-import { useEffect, useMemo } from 'react'
+import { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { setLocalisation } from '../../actions/actions'
-import frenchCountryPrepositions from './frenchCountryPrepositions.yaml'
-import supportedCountriesYAML from './supportedCountries.yaml'
+import useBranchData from '../useBranchData'
 
 const API = '/geolocation'
-
-export type Region = {
-	PR: string
-	nom: string
-	code: string
-	gentilé: string
-	drapeau: string
-}
-
-export const supportedRegions: Region[] = supportedCountriesYAML
 
 // Other alternatives :
 // https://positionstack.com/product
@@ -72,63 +61,4 @@ export default () => {
 	}, [localisation])
 
 	return localisation
-}
-
-export const useFlag = (localisation) => {
-	const flag = useMemo(() => {
-		if (localisation?.country?.code != null) {
-			if (isSupportedRegion(localisation?.country?.code)) {
-				return getSupportedFlag(localisation?.country.code)
-			} else {
-				return getFlagImgSrc('FR')
-			}
-		}
-	}, [localisation])
-
-	return flag
-}
-
-export const getFlagImgSrc = (code) =>
-	//	code && `https://flagcdn.com/96x72/${code.toLowerCase()}.png`
-	//	was down 27/09
-	code &&
-	`https://cdn.jsdelivr.net/npm/svg-country-flags@1.2.10/svg/${code.toLowerCase()}.svg`
-
-export const getCountryNameInFrench = (code) => {
-	// For now, website is only available in French, this function enables to adapt message
-	// for French Language according to the country detected.
-	// Including French prepositions subtelties.
-	if (!code) {
-		return undefined
-	}
-
-	const regionNamesInFrench = new Intl.DisplayNames(['fr'], { type: 'region' }),
-		countryNameAuto = regionNamesInFrench.of(code),
-		countryName =
-			countryNameAuto === 'France' ? 'France métropolitaine' : countryNameAuto,
-		preposition = (countryName && frenchCountryPrepositions[countryName]) || ''
-
-	return `${preposition} ${countryName}`
-}
-
-export const getSupportedFlag = (inputCode) => {
-	if (!inputCode) return
-
-	const supported = supportedRegions.find((c) => c.code === inputCode)
-
-	const code = supported?.drapeau || inputCode
-
-	return getFlagImgSrc(code)
-}
-
-export const isSupportedRegion = (inputCode) => {
-	if (!inputCode) {
-		return undefined
-	}
-
-	const supported = supportedRegions.find((c) => c.code === inputCode)
-	if (supported?.inactif === 'oui') {
-		return false
-	}
-	return supported
 }
